lib: move Concentrator type to its own module
